feat: cleam up code, change file name textarea on biotextarea
